2015 updates from 2016, pushed to 2017

+++ src/od-linux/memory.c
@@ -14,9 +14,9 @@
 #if defined(PPC_CPU)
 void ppc_cacheflush(void* start, int length)
 {
-  uintptr_t stop = ((uintptr_t)start + length);
+  intptr_t stop = ((intptr_t)start + (size_t)length);
 
-  flush_cache_range((uintptr_t)start, stop);
+  flush_cache_range((intptr_t)start, stop);
 }
 #endif
